requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-index-url --no-emit-trusted-host
#
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via jsonschema
django==4.0.2
    # via
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==3.11.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.21.2
    # via -r requirements.in
environs==9.5.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.4.0
    # via drf-spectacular
marshmallow==3.14.1
    # via environs
psycopg2==2.9.3
    # via -r requirements.in
pyrsistent==0.18.1
    # via jsonschema
python-dotenv==0.19.2
    # via environs
pytz==2021.3
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via drf-spectacular

# The following packages are considered to be unsafe in a requirements file:
# setuptools
